moved websocket to same pattern as ozw also required moving the bridge schema to a separate file to avoid circular dependencies
